requirements.txt
pytest==8.1.1
requests==2.31.0
allure-pytest==2.13.5
python-dotenv==1.0.1
